requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
httpx==0.24.1
python-telegram-bot==20.5
openai==1.12.0
gTTS==2.4.0
python-dotenv==1.0.0
aioredis==2.0.1 
